Make the table scroll with fixed header and height responsive
We replace the pure css approach by the jquery plugin
https://github.com/alalonde/angular-scrollable-table

Still need to make it width responsive when increasing the window width.

scrollable-table uses an isolate scope in which it puts the filtered
list, which cannot be accessed from the main scope.

Two solutions:

1. Make a function to copy everything back to the main scope
2. Use controllerAS to create a "namespace" linked to our controller and
copy the results.

I chose solution 2.