Upgrade the local db
The upgrade function loops over all the activities returned by Strava
and checks whether it is already in the local db. Moreover, it updates
the name, description and red_points (if equals 0) fields if needed.

Upgrading a whole db takes a while as each activity requires at least
one request to the Strava API (description field) and maybe an other one
for the red_points.